// src/modules/loader/dtos/load-account.dto.ts
export class LoadAccountDto {
  ACCOUNTNO: string;
  ACCTSTATUSCODE: string;
  ACCTTYPE: string;
  APPRAISALTYPE: string;
  DEFAULTTAXDISTRICT: string;
  PARCELNO: string;
  PREDIRECTION: string;
  PROPERTYCITY: string;
  STREETNAME: string;
  STREETNO: string;
  STREETTYPE: string;
  ACTUALVALUE: number;
  ASSESSEDVALUE: number;
  MILLLEVY: number;
  TAXYEAR: number;
  TOTALTAXDOLLARS: number;
  APPRAISALTYPE1: string;
}