// src/modules/loader/dtos/load-value-details.dto.ts
export class LoadValueDetailsDto {
  ACCOUNTNO: string;
  ACCTSTATUSCODE: string;
  APPRAISALTYPE: string;
  PARCELNO: string;
  ABSTRACTCODE: string;
  ABSTRACTDESCRIPTION: string;
  ACTUALVALUE: number;
  ASSESSEDVALUE: number;
  LANDACRES: number;
  LANDSF: number;
  TAXDISTRICT: string;
  APPRAISALTYPE1: string;
}